Fix watch_dir path mix up
